Refactor: new widget UsedForList
